Extract LocalManifest to its own module
